Update Vagrant, add support services
This change reintroduces the usage of Vagrant for windows users.
Additionally, the logs have been improved for better understanding of
which stage is the bootstrap at.
Finally, additional services have been added such as portainer, to check
the status of the containers, and drone, a CI pipeline. Clearly the API
keys for drone have been revoked so for now it is not really usable, but
it is a WIP