Add script for automatic deployment on cron job
While in the future an approach hook-based will work better (since
state-of-the-art) for now this will do. The deploy script is meant to be
run with privileges (to access the docker daemon) and to be scheduled as
a cron job: there will be a check if updates of upstream and a pull
with update of submodules, and a restart of the services (docker-compose
will take care of restarting only the changed containers).
There are also other options to 'refresh' the installation by deleting
the .env file, or to nuke it and re-downloading it from scratch.