Add empty acme.json for no-hassle startups
The empty file is generated with the correct 600 permission into secrets
by the oms.sh script. It is generated in the secrets folder to avoid
accidental upload of the certificates.
The file must be mounted into the traefik container for persistence, so
no "secrets" option from docker-compose manifest can be used