(Impr) introduce basic security concepts around
In an attempt to standardise some content (MEMB-180 & MEMB-104) we put
some passwords as single default ones. The rationale is that if we have
a system that can login with 5ecr3t, we forgot to set something -> we
are idiots.

So, it's a 'standard' for the new passwords, in short :p

Also some routes are behind (basic) auth by traefik